// angular-photo-gallery/src/app/interfaces/Photo.ts
export interface Photo {
    _id?: string;
    title: String;
    description: String;
    imagePath: string;

    nombre :String;
    apellido: String;  
    email:String;
    fecha_nacimiento: String;
    
}
// export interface User {
//     _id?: string;  
//     correo:String;
//     contraseña:String;

// }